# apigateway/openapi.yaml
openapi: 3.1.0
info:
  title: E-commerce API Gateway
  version: 1.1.0
  description: Consolidated OpenAPI spec for the API Gateway proxying User, Product, and Order services.
servers:
  - url: http://localhost:8083
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      description: Prevent duplicate order creation. If provided, repeated requests return the same order.
      schema:
        type: string
        maxLength: 64
  schemas:
    Error:
      type: object
      properties:
        message: { type: string }
        code: { type: string }
        details: { type: object, additionalProperties: true }
...
    Address:
      type: object
      properties:
        id: { type: string }
        line1: { type: string }
        line2: { type: string, nullable: true }
        city: { type: string }
        state: { type: string }
        postal_code: { type: string }
        country: { type: string }
        phone: { type: string, nullable: true }
        is_default: { type: boolean }
...
    User:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string }
        phone: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
...
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        price: { type: number, format: float }
        stock: { type: integer }
...
    OrderStatus:
      type: string
      enum: [PENDING, PAID, CANCELLED]
    OrderItem:
      type: object
      properties:
        product_id: { type: string }
        quantity: { type: integer }
        price: { type: number }
    Order:
      type: object
      properties:
        id: { type: string }
        user_id: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        total_amount: { type: number }
        shipping_address_id: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    CreateOrderItem:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: string }
        quantity: { type: integer, minimum: 1 }
    CreateOrderRequest:
      type: object
      required: [userId, items]
      properties:
        userId: { type: string }
        items:
          type: array
          items:
            $ref: '#/components/schemas/CreateOrderItem'
        shippingAddressId: { type: string, nullable: true }
    OrdersListResponse:
      type: object
      properties:
        orders:
          type: array
          items: { $ref: '#/components/schemas/Order' }
        nextCursor: { type: string, nullable: true }
    OrderDetails:
      type: object
      properties:
        id: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        total_amount: { type: number }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        userId: { type: string }
        shippingAddressId: { type: string, nullable: true }
        shippingAddress:
          oneOf:
            - $ref: '#/components/schemas/Address'
            - type: 'null'
        user:
          oneOf:
            - $ref: '#/components/schemas/User'
            - type: 'null'
        items:
          type: array
          items:
            type: object
            properties:
              productId: { type: string }
              quantity: { type: integer }
              product:
                oneOf:
                  - $ref: '#/components/schemas/Product'
                  - type: 'null'
security:
  - bearerAuth: []
paths:
...
  /api/v1/orders:
    post:
      tags: [Orders]
      summary: Create order
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrderRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  status: { $ref: '#/components/schemas/OrderStatus' }
        '200': { description: Returned existing order for idempotent request }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503': { description: Dependency unavailable }
    get:
      tags: [Orders]
      summary: List orders
      parameters:
        - in: query
          name: userId
          schema: { type: string }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/OrderStatus' }
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrdersListResponse' }
  /api/v1/orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order by ID
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/orders/{orderId}/details:
    get:
      tags: [Orders]
      summary: Get order with enriched user and product details
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderDetails' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/orders/{orderId}/pay:
    post:
      tags: [Orders]
      summary: Mark order paid
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Paid }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Conflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/orders/{orderId}/cancel:
    post:
      tags: [Orders]
      summary: Cancel order
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Cancelled }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Conflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }